Add ActivitySpots to the SettlementMap tool
Spots are filtered by the Function Type to reduce label overload on the map.
In addition, it continues the rework/consolidation of the MapLayers

Changes:
- Seperate StructureMapLayer into 2 seperate layers for Construction & Building
- AbstractMapLayer provides common method to draw SVG & Rectangles onto the map
- VehicleMapLayer now used the SVG method of the parent AbstractMapLayer
- BuildingMapLayer drops the ability to draw one building as it is never used
- BuildingMapLayer will draw Activity Spots on the map with their name filtered according to FunctionType
- SettlementMapPanel holds the displayed FunctionTypes for the Spots
- SettlementTransparentPanel adds a Spots sub-menu to the popup label menu; only the FunctionType having spots are shown
- BuildingConfig now identifies FunctionTypes that use ActivitySpots
- Drop the unused Building parameter from the SettlementMapLayer.drawLayer method and update implementations
- Console bed command changed to list the Beds and the assigned Person instead of counts
- Changed the assignments of beds in first pulse to ignore guest beds to reduce overallocation
- Generally removed CodeSmells where possible and in scope
- The Single Scenario configuration has too many persons for the number of beds so reduce

Part of #1166 #1143
