+++ mars-sim-ui/src/main/java/com/mars_sim/ui/swing/tool/settlement/AbstractMapLayer.java
@@ -21,6 +21,9 @@ import java.awt.image.BufferedImage;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.batik.gvt.GraphicsNode;
+
+import com.mars_sim.mapdata.location.LocalBoundedObject;
 import com.mars_sim.mapdata.location.LocalPosition;
 
 /**
@@ -29,8 +32,20 @@ import com.mars_sim.mapdata.location.LocalPosition;
 public abstract class AbstractMapLayer implements SettlementMapLayer {
     // A data record to represent a color choice.
     protected record ColorChoice(Color text, Color outline) {}
+	
+    // A data record to represent a structure key.
+    private record StructureKey(GraphicsNode svg, double width, double length) {}
+
+	private static final float[] DASHES = {50.0f, 20.0f, 10.0f, 20.0f};
+    
+    // See https://docstore.mik.ua/orelly/java-ent/jfc/ch04_05.htm for instructions on BasicStroke
+    private static final BasicStroke THIN_DASH = new BasicStroke(2.0f,
+    	      BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER, 10.0f, DASHES, 0.0f);
+	private static final BasicStroke THICK_DASH = new BasicStroke(10.0f,
+			  BasicStroke.CAP_ROUND, BasicStroke.JOIN_ROUND, 50.0f, DASHES, 0.0f);
 
     private Map<String, BufferedImage> labelImageCache = new HashMap<>();
+	private Map<Double, Map<StructureKey, BufferedImage>> svgImageCache = new HashMap<>();
 
     /**
 	 * Draw an oval at a settlement.
@@ -283,4 +298,244 @@ public abstract class AbstractMapLayer implements SettlementMapLayer {
 
 		return bufferedImage;
 	}
+
+	/**
+     * Gets a buffered image for a given graphics node.
+     * 
+     * @param svg the SVG graphics node.
+     * @param width the structure width (meters).
+     * @param length the structure length (meters).
+     * @param patternSVG the pattern SVG graphics node (null if no pattern).
+     * @return buffered image.
+     */
+    protected BufferedImage getBufferedImage(
+            GraphicsNode svg, double width, double length,
+            GraphicsNode patternSVG, double scale) {
+
+        // Get image cache for current scale or create it if it doesn't exist.
+        Map<StructureKey, BufferedImage> imageCache = null;
+        if (svgImageCache.containsKey(scale)) {
+            imageCache = svgImageCache.get(scale);
+        }
+        else {
+            imageCache = new HashMap<>(100);
+            svgImageCache.put(scale, imageCache);
+        }
+
+        // Get image from image cache or create it if it doesn't exist.
+        BufferedImage image = null;
+        StructureKey buildingKey = new StructureKey(svg, width, length);
+        if (imageCache.containsKey(buildingKey)) {
+            image = imageCache.get(buildingKey);
+        }
+        else {
+            image = createBufferedImage(svg, width, length, patternSVG, scale);
+            imageCache.put(buildingKey, image);
+        }
+
+        return image;
+    }
+
+    /**
+     * Creates a buffered image from a SVG graphics node.
+     * 
+     * @param svg the SVG graphics node.
+     * @param width the structure width (meters).
+     * @param length the structure length (meters).
+     * @param patternSVG the pattern SVG graphics node (null if no pattern).
+     * @return the created buffered image.
+     */
+    private BufferedImage createBufferedImage(GraphicsNode svg, double width, double length,
+            GraphicsNode patternSVG, double scale) {
+
+        int imageWidth = (int) (width * scale);
+        if (imageWidth <= 0) {
+            imageWidth = 1;
+        }
+        int imageLength = (int) (length * scale);
+        if (imageLength <= 0) {
+            imageLength = 1;
+        }
+        BufferedImage bufferedImage = new BufferedImage(
+                imageWidth, imageLength,
+                BufferedImage.TYPE_INT_ARGB
+                );
+
+        // Determine bounds.
+        Rectangle2D bounds = svg.getBounds();
+
+        // Determine transform information.
+        double scalingWidth = width / bounds.getWidth() * scale;
+        double scalingLength = length / bounds.getHeight() * scale;
+
+        // Draw the SVG image on the buffered image.
+        Graphics2D g2d = (Graphics2D) bufferedImage.getGraphics();
+        g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
+        svg.setTransform(AffineTransform.getScaleInstance(scalingWidth, scalingLength));
+        svg.paint(g2d);
+
+        // Draw repeating pattern SVG image on the buffered image.
+        if (patternSVG != null) {
+            double patternScaling;
+            double patternWidth;
+            double patternLength;
+
+            double originalProportions = bounds.getWidth() / bounds.getHeight();
+            double finalProportions = width / length;
+            Rectangle2D patternBounds = patternSVG.getBounds();
+            if ((finalProportions / originalProportions) >= 1D) {
+                patternScaling = scalingLength;
+                patternLength = length * (patternBounds.getHeight() / bounds.getHeight());
+                patternWidth = patternLength * (patternBounds.getWidth() / patternBounds.getHeight());
+            }
+            else {
+                patternScaling = scalingWidth;
+                patternWidth = width * (patternBounds.getWidth() / bounds.getWidth());
+                patternLength = patternWidth * (patternBounds.getHeight() / patternBounds.getWidth());
+            }
+
+            AffineTransform patternTransform = new AffineTransform();
+            patternTransform.scale(patternScaling, patternScaling);
+            for (double x = 0D; x < length; x += patternLength) {
+                patternTransform.translate(0D, x * bounds.getHeight());
+                double y = 0D;
+                for (; y < width; y += patternWidth) {
+                    patternTransform.translate(y * bounds.getWidth(), 0D);
+                    patternSVG.setTransform(patternTransform);
+                    patternSVG.paint(g2d);
+                    patternTransform.translate(y * bounds.getWidth() * -1D, 0D);
+                }
+                patternTransform.translate(0D, x * bounds.getHeight() * -1D);
+            }
+        }
+
+        // Cleanup and return image
+        g2d.dispose();
+
+        return bufferedImage;
+    }
+
+	/**
+     * Draws a structure using SVG on the map.
+     * 
+     * @param scale Scale
+     * @param placement Placement of structure
+     * @param svg the SVG graphics node.
+     * @param patternSVG the pattern SVG graphics node (null if no pattern).
+	 * @param selectedColor the color to draw as selected
+     */
+    protected void drawStructure(Graphics2D g2d, double scale, LocalBoundedObject placement,
+            GraphicsNode svg, GraphicsNode patternSVG, Color selectedColor) {
+
+        double xLoc = placement.getXLocation();
+        double yLoc = placement.getYLocation();
+        double width = placement.getWidth();
+        double length = placement.getLength();
+        double facing = placement.getFacing();
+
+        // Save original graphics transforms.
+        AffineTransform saveTransform = g2d.getTransform();
+   
+        // Determine bounds.
+        Rectangle2D bounds = svg.getBounds();
+
+        // Determine transform information.
+        double scalingWidth = width / bounds.getWidth() * scale;
+        double scalingLength = length / bounds.getHeight() * scale;
+        double boundsPosX = bounds.getX() * scalingWidth;
+        double boundsPosY = bounds.getY() * scalingLength;
+        double centerX = width * scale / 2D;
+        double centerY = length * scale / 2D;
+        double translationX = (-1D * xLoc * scale) - centerX - boundsPosX;
+        double translationY = (-1D * yLoc * scale) - centerY - boundsPosY;
+        double facingRadian = facing / 180D * Math.PI;
+  
+        AffineTransform newTransform = new AffineTransform();
+        
+		// Draw buffered image of structure.
+		BufferedImage image = getBufferedImage(svg, width, length, patternSVG, scale);
+		if (image != null) {
+			
+			// Apply graphic transforms for structure.		
+			newTransform.translate(translationX, translationY);
+			newTransform.rotate(facingRadian, centerX + boundsPosX, centerY + boundsPosY);
+			
+			g2d.transform(newTransform);
+			g2d.drawImage(image, 0, 0, null);
+		}
+
+        if (selectedColor != null) {   
+            AffineTransform newTransform1 = new AffineTransform();
+        	newTransform1.scale(scalingWidth, scalingLength);
+            g2d.transform(newTransform1);
+            
+			// Draw the dashed border over the selected building
+			g2d.setPaint(selectedColor);
+
+			// Save original stroke
+        	Stroke oldStroke = g2d.getStroke();
+			g2d.setStroke(THICK_DASH);                                           
+			g2d.draw(bounds);
+			
+			// Restore the stroke
+			g2d.setStroke(oldStroke);
+        }
+        
+        // Restore original graphic transforms.
+        g2d.setTransform(saveTransform);
+    }
+
+	/**
+     * Draws a rectangle on the map.
+     * 
+     * @param scale Scale
+     * @param placement Placement of structure
+     * @param color the color to display the rectangle if no SVG image.
+     */
+    protected void drawRectangle(
+            Graphics2D g2d, double scale, LocalBoundedObject placement,
+            Color color, Color selectedColor) {
+
+        double width = placement.getWidth();
+        double length = placement.getLength();
+
+        // Save original graphics transforms.
+        AffineTransform saveTransform = g2d.getTransform();
+        // Save original stroke
+        Stroke oldStroke = g2d.getStroke();
+   
+        // Determine bounds.
+        Rectangle2D bounds = new Rectangle2D.Double(0, 0, width, length);
+
+        // Determine transform information.
+        double scalingWidth = width / bounds.getWidth() * scale;
+        double scalingLength = length / bounds.getHeight() * scale;
+  
+        AffineTransform newTransform = new AffineTransform();
+       
+		// Draw filled rectangle.
+		newTransform.scale(scalingWidth, scalingLength);
+		g2d.transform(newTransform);
+		
+		g2d.setColor(color);
+		g2d.fill(bounds);
+		
+		if (selectedColor != null) {
+			// Draw the dashed border
+			g2d.setPaint(selectedColor);
+			g2d.setStroke(THIN_DASH);
+			g2d.draw(bounds);
+			g2d.setStroke(oldStroke);
+		}
+        
+        // Restore original graphic transforms.
+        g2d.setTransform(saveTransform);
+    }
+
+	
+    @Override
+    public void destroy() {
+		labelImageCache.clear();
+        svgImageCache.clear();
+    }   
 }
