+++ mars-sim-ui/src/main/java/com/mars_sim/ui/swing/tool/settlement/RobotMapLayer.java
@@ -13,7 +13,6 @@ import java.util.List;
 import com.mars_sim.core.CollectionUtils;
 import com.mars_sim.core.robot.Robot;
 import com.mars_sim.core.structure.Settlement;
-import com.mars_sim.core.structure.building.Building;
 
 /**
  * A settlement map layer for displaying Robots.
@@ -38,7 +37,7 @@ public class RobotMapLayer extends WorkerMapLayer<Robot> {
 
 	@Override
 	public void displayLayer(
-		Graphics2D g2d, Settlement settlement, Building building,
+		Graphics2D g2d, Settlement settlement,
 		double xPos, double yPos, int mapWidth, int mapHeight,
 		double rotation, double scale) {
